Move impl to single file
